refactor: key extraction
